+++ R/spct.classes.r
@@ -14,7 +14,8 @@ spct_classes <- function() {
     "raw_spct", "cps_spct",
     "filter_spct", "reflector_spct",
     "source_spct", "object_spct",
-    "response_spct", "chroma_spct", "generic_spct")
+    "response_spct", "chroma_spct",
+    "solute_spct", "generic_spct")
 }
 
 # check -------------------------------------------------------------------
@@ -396,6 +397,67 @@ check_spct.filter_spct <-
       if (exists("A", x, mode = "numeric", inherits = FALSE) && anyNA(x[["A"]])) {
         warning("At least one NA in 'A'")
       }
+    }
+    x
+  }
+
+#' @describeIn check_spct Specialization for solute_spct.
+#' @export
+check_spct.solute_spct <-
+  function(x,
+           byref = TRUE,
+           strict.range = getOption("photobiology.strict.range", default = FALSE),
+           force = FALSE,
+           multiple.wl = getMultipleWl(x),
+           ...) {
+
+    range_check <- function(x, strict.range, k.base) {
+      K.col <- paste("K", k.base, sep = ".")
+      stopifnot(K.col %in% names(x))
+
+      if (!all(is.na(x[[K.col]]))) {
+        K.col.min <- min(x[[K.col]], na.rm = TRUE)
+        K.col.max <- max(x[[K.col]], na.rm = TRUE)
+        if (K.col.min < -1e-4) {
+          message.text <-
+            paste0(
+              "Negative attenuation values [",
+              formatted_range(c(K.col.min, K.col.max)),
+              "]",
+              sep = ""
+            )
+          if (is.null(strict.range) || is.na(strict.range)) {
+            message(message.text)
+          } else if (strict.range) {
+            stop(message.text)
+          } else if (!strict.range) {
+            warning(message.text)
+          } else {
+            stop ("Bad argument for 'strict.range': ", strict.range)
+          }
+        }
+      }
+    }
+
+    x <- check_spct.generic_spct(x, multiple.wl = multiple.wl)
+    if (is.null(getKType(x))) {
+      setKType(x, "attenaution")
+      warning("Missing K.type attribute replaced by 'attenuation'")
+    }
+
+    if (exists("K.mole", x, mode = "numeric", inherits = FALSE)) {
+      range_check(x, strict.range = strict.range, k.base = "mole")
+      if (getOption("photobiology.verbose") && anyNA(x[["K.mole"]])) {
+        warning("At least one NA in 'K.mole'")
+      }
+    } else if (exists("K.mass", x, mode = "numeric", inherits = FALSE)) {
+      range_check(x, strict.range = strict.range, k.base = "mass")
+      if (getOption("photobiology.verbose") && anyNA(x[["K.mass"]])) {
+        warning("At least one NA in 'K.mass'")
+      }
+    } else {
+      warning("No coefficient of attenuation data found in solute_spct")
+      x[["K.mole"]] <- NA_real_
     }
     x
   }
@@ -1004,6 +1066,46 @@ setFilterSpct <-
     invisible(x)
   }
 
+#' @describeIn setGenericSpct Set class of an object to "solute_spct".
+#'
+#' @param K.type character One of "attenuation", "absorption", or "scattering".
+#' @param strict.range logical Flag indicating whether off-range values result in an
+#'   error instead of a warning.
+#' @export
+#'
+#'
+#' @note For non-diffusing materials like glass an approximate \code{Rfr.constant}
+#'   value can be used to interconvert "total" and "internal" transmittance
+#'   values. Use \code{NA} if not known, or not applicable, e.g., for materials
+#'   subject to internal scattering.
+#'
+setSoluteSpct <-
+  function(x,
+           K.type = c("attenuation", "absorption", "scattering"),
+           strict.range = getOption("photobiology.strict.range", default = FALSE),
+           multiple.wl = 1L,
+           idfactor = NULL) {
+    name <- substitute(x)
+    if (is.solute_spct(x) &&
+        getKType(x) != "unknown") {
+      if (length(K.type) > 1) {
+        K.type <- getKType(x)
+      } else if (K.type != getKType(x)) {
+        warning("Overwriting attribute 'K.type' from ", getKType(x),
+                " into ", K.type)
+      }
+    }
+    setGenericSpct(x, multiple.wl = multiple.wl, idfactor = idfactor)
+    class(x) <- c("solute_spct", class(x))
+    setKType(x, K.type[1])
+    x <- check_spct(x, strict.range = strict.range)
+    if (is.name(name)) {
+      name <- as.character(name)
+      assign(name, x, parent.frame(), inherits = TRUE)
+    }
+    invisible(x)
+  }
+
 #' @describeIn setGenericSpct Set class of a an object to "reflector_spct".
 #'
 #' @param Rfr.type character A string, either "total" or "specular".
@@ -1012,7 +1114,7 @@ setFilterSpct <-
 #'
 setReflectorSpct <-
   function(x,
-           Rfr.type=c("total", "specular"),
+           Rfr.type = c("total", "specular"),
            strict.range = getOption("photobiology.strict.range", default = FALSE),
            multiple.wl = 1L,
            idfactor = NULL) {
@@ -1050,9 +1152,6 @@ setObjectSpct <-
            multiple.wl = 1L,
            idfactor = NULL) {
     name <- substitute(x)
-    if (Tfr.type == "total" && Rfr.type != "total") {
-      warning("Rfr is not \"total\", making conversions between Afr and Tfr impossible.")
-    }
     if ((is.filter_spct(x) || is.object_spct(x)) && getTfrType(x) != "unknown") {
       if (length(Tfr.type) > 1) {
         Tfr.type <- getTfrType(x)
@@ -1060,6 +1159,8 @@ setObjectSpct <-
         warning("Overwriting attribute 'Tfr.type' from ", getTfrType(x),
                 " into ", Tfr.type)
       }
+    } else {
+      Tfr.type <- Tfr.type[1]
     }
     if ((is.reflector_spct(x) || is.object_spct(x)) && getRfrType(x) != "unknown") {
       if (length(Rfr.type) > 1) {
@@ -1068,6 +1169,11 @@ setObjectSpct <-
         warning("Overwriting attribute 'Rfr.type' from ", getRfrType(x),
                 " into ", Rfr.type)
       }
+    } else {
+      Rfr.type <- Rfr.type[1]
+    }
+    if (Tfr.type == "total" && Rfr.type != "total") {
+      message("Rfr is not \"total\", making conversions between Afr and Tfr impossible.")
     }
     setGenericSpct(x, multiple.wl = multiple.wl, idfactor = idfactor)
     class(x) <- c("object_spct", class(x))
@@ -1228,6 +1334,11 @@ is.reflector_spct <- function(x) inherits(x, "reflector_spct")
 #' @export
 #'
 is.object_spct <- function(x) inherits(x, "object_spct")
+
+#' @rdname is.generic_spct
+#' @export
+#'
+is.solute_spct <- function(x) inherits(x, "solute_spct")
 
 #' @rdname is.generic_spct
 #' @export
@@ -1425,6 +1536,41 @@ is_transmittance_based <- function(x) {
   }
 }
 
+# is_mole_based ---------------------------------------------------------
+
+#' Query if a spectrum contains mole or mass based data
+#'
+#' Functions to check if an solute attenuation spectrum contains coefficients
+#' on expressed on mole of mass base.
+#'
+#' @param x an R object
+#'
+#' @return \code{is_mole_based} returns TRUE if its argument is a
+#'   \code{solute_spct} object that contains spectral \code{K.mole} data and
+#'   \code{FALSE} if it contains \code{K.mass} data, but returns NA for any
+#'   other R object, including those belonging other \code{generic_spct}-derived
+#'   classes. \code{is_mass_based} returns the complement of
+#'   \code{is_mole_based}.
+#'
+#' @export
+#' @family query units functions
+#'
+#' @examples
+#' print("missing example")
+#'
+is_mole_based <- function(x) {
+  if (is.solute_spct(x) || is.summary_solute_spct(x)) {
+    return("K.mole" %in% names(x))
+  } else {
+    return(NA_integer_)
+  }
+}
+
+#' @rdname is_mole_based
+#'
+is_mass_based <- function(x) {
+  !is_mole_based(x)
+}
 
 # time.unit attribute -----------------------------------------------------
 
@@ -1977,6 +2123,81 @@ getRfrType <- function(x) {
       Rfr.type <- "unknown"
     }
     return(Rfr.type[[1]])
+  } else {
+    return(NA_character_)
+  }
+}
+
+# Tfr.type attribute ------------------------------------------------------
+
+#' Set the "K.type" attribute
+#'
+#' Function to set by reference the "K.type" attribute of an existing
+#' solute_spct object
+#'
+#' @param x a solute_spct or a summary_solute_spct object.
+#' @param K.type a character string, either "attenuation", "absorption" or "scattering".
+#'
+#' @return x
+#' @note This function alters x itself by reference and in addition
+#'   returns x invisibly. If x is not a solute_spct object, x is not modified
+#'   The behaviour of this function is 'unusual' in that the default for
+#'   parameter \code{K.type} is used only if \code{x} does not already have
+#'   this attribute set.
+#'
+#' @export
+#' @family K attribute functions
+#' @examples
+#' print("missing example")
+#'
+setKType <- function(x,
+                     K.type = c("attenuation", "absorption", "scattering")) {
+  name <- substitute(x)
+  if (length(K.type) > 1) {
+    if (getKType(x) != "unknown") {
+      K.type <- getKType(x)
+    } else {
+      K.type <- K.type[[1]]
+    }
+  }
+  if (is.solute_spct(x) || is.summary_solute_spct(x)) {
+    if  (!(K.type %in% c("attenuation", "absorption", "scattering"))) {
+      warning("Invalid 'K.type' argument, only 'attenuation', 'absorption' and 'scattering' supported.")
+      return(x)
+    }
+    attr(x, "K.type") <- K.type
+    if (is.name(name)) {
+      name <- as.character(name)
+      assign(name, x, parent.frame(), inherits = TRUE)
+    }
+  }
+  invisible(x)
+}
+
+#' Get the "K.type" attribute
+#'
+#' Function to read the "K.type" attribute of an existing solute_spct object.
+#'
+#' @param x a solute_spct object
+#'
+#' @return character string
+#'
+#' @note If x is not a \code{solute_spct} or a \code{summary_solute_spct} object,
+#'   \code{NA} is returned.
+#'
+#' @export
+#' @family K attribute functions
+#' @examples
+#' print("missing example")
+#'
+getKType <- function(x) {
+  if (is.solute_spct(x) || is.summary_solute_spct(x)) {
+    K.type <- attr(x, "K.type", exact = TRUE)
+    if (is.null(K.type) || is.na(K.type)) {
+      # need to handle corrupted objects
+      K.type <- "unknown"
+    }
+    return(K.type[[1]])
   } else {
     return(NA_character_)
   }
@@ -2564,7 +2785,7 @@ convertTfrType <- function(x, Tfr.type = NULL) {
     return(invisible(x))
   }
 
-  if (is.null(Tfr.type) || Tfr.type == getTfrType(x)) {
+  if (is.null(Tfr.type) || Tfr.type[1] == getTfrType(x)[1]) {
     # nothing to do
     return(invisible(x))
   }